VideoMaker/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
